feat: latest searches component in users page (#38)
close #38
