+++ src/app/users/components/Sections/index.ts
@@ -1,5 +1,5 @@
 export * from './ActionsSection'
 export * from './FiltersSection'
-export * from './LatestSearchesSection'
+export * from './LatestSearchesSection/LatestSearchesSection'
 export * from './ResultsSection'
